requirements.txt
# LayerLearn — Dependencies
# ========================

# LLM backend
ollama>=0.4.0

# Voice pipeline
faster-whisper>=1.0.0
edge-tts>=6.1.0
sounddevice>=0.4.6
pydub>=0.25.1

# Screen capture
mss>=9.0.0
pillow>=10.0.0

# Input / clipboard
pynput>=1.7.6
pyperclip>=1.8.2

# Terminal UI
rich>=13.0.0

# Logging
loguru>=0.7.0

# Config
python-dotenv>=1.0.0

# Numeric
numpy>=1.24.0

# System monitoring
psutil>=5.9.0